HDDS-9520. Merge some TestOzoneManagerHA subclasses
