--- hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/ozone/om/TestOzoneManagerHAMetrics.java
@@ -1,84 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with this
- * work for additional information regarding copyright ownership.  The ASF
- * licenses this file to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package org.apache.hadoop.ozone.om;
-
-import org.apache.hadoop.ozone.om.ha.OMHAMetrics;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-import java.util.List;
-
-/**
- * Test Ozone Manager HA Metrics.
- */
-public class TestOzoneManagerHAMetrics extends TestOzoneManagerHA {
-
-  @Test
-  public void testOMHAMetrics() throws Exception {
-    // Get leader OM
-    OzoneManager leaderOM = getCluster().getOMLeader();
-    // Store current leader's node ID,
-    // to use it after restarting the OM
-    String leaderOMId = leaderOM.getOMNodeId();
-    // Get a list of all OMs
-    List<OzoneManager> omList = getCluster().getOzoneManagersList();
-    // Check metrics for all OMs
-    checkOMHAMetricsForAllOMs(omList, leaderOMId);
-
-    // Restart leader OM
-    getCluster().shutdownOzoneManager(leaderOM);
-    getCluster().restartOzoneManager(leaderOM, true);
-    waitForLeaderToBeReady();
-
-    // Get the new leader
-    OzoneManager newLeaderOM = getCluster().getOMLeader();
-    String newLeaderOMId = newLeaderOM.getOMNodeId();
-    // Get a list of all OMs again
-    omList = getCluster().getOzoneManagersList();
-    // New state for the old leader
-    int newState = leaderOMId.equals(newLeaderOMId) ? 1 : 0;
-
-    // Get old leader
-    OzoneManager oldLeader = getCluster().getOzoneManager(leaderOMId);
-    // Get old leader's metrics
-    OMHAMetrics omhaMetrics = oldLeader.getOmhaMetrics();
-
-    Assertions.assertEquals(newState,
-        omhaMetrics.getOmhaInfoOzoneManagerHALeaderState());
-
-    // Check that metrics for all OMs have been updated
-    checkOMHAMetricsForAllOMs(omList, newLeaderOMId);
-  }
-
-  private void checkOMHAMetricsForAllOMs(List<OzoneManager> omList,
-                                         String leaderOMId) {
-    for (OzoneManager om : omList) {
-      // Get OMHAMetrics for the current OM
-      OMHAMetrics omhaMetrics = om.getOmhaMetrics();
-      String nodeId = om.getOMNodeId();
-
-      // If current OM is leader, state should be 1
-      int expectedState = nodeId
-          .equals(leaderOMId) ? 1 : 0;
-      Assertions.assertEquals(expectedState,
-          omhaMetrics.getOmhaInfoOzoneManagerHALeaderState());
-
-      Assertions.assertEquals(nodeId, omhaMetrics.getOmhaInfoNodeId());
-    }
-  }
-
-}
